meiwen/requirement.txt
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-import-export==1.0.1
django-reversion==2.0.13
djangorestframework==3.8.2
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.16.0
httpie==0.9.9
httplib2==0.11.3
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.0.0
Pygments==2.2.0
PyMySQL==0.8.1
pytz==2018.4
PyYAML==3.12
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
